Progress

+++ octopoes/octopoes/core/service.py
@@ -402,8 +402,6 @@ class OctopoesService:
             refs = set(origin.result)
             origin.phantom_result = self.ooi_repository.load_bulk_as_list(refs, valid_time)
             origin.result = []
-            for ref in refs:
-                self.ooi_repository.delete(ref, valid_time)
             self.origin_repository.save(origin, valid_time)
 
     def nibblet_dephantomize_result(self, origin: Origin, valid_time: datetime) -> None:
@@ -414,8 +412,6 @@ class OctopoesService:
             and not origin.result
         ):
             origin.result = [result.reference for result in origin.phantom_result]
-            for ooi in origin.phantom_result:
-                self.ooi_repository.save(ooi, valid_time)
             origin.phantom_result = []
             self.origin_repository.save(origin, valid_time)
 
@@ -548,6 +544,12 @@ class OctopoesService:
 
         dereferenced_oois = event.old_data - event.new_data
         for reference in dereferenced_oois:
+            if (
+                event.new_data.origin_type == OriginType.NIBBLET
+                and event.new_data.phantom_result
+                and any(reference == ooi.reference for ooi in event.new_data.phantom_result)
+            ):
+                continue
             self._delete_ooi(reference, event.valid_time)
 
     def _on_delete_origin(self, event: OriginDBEvent) -> None:
